implement `number` and treat empty cells as 0
`number` takes a one or more arguments and returns the first argument
that is a finite, non-NaN number, or 0 if there is no such argument.

Previously, empty cells were treated as the identify for the ongoing
operation. This meant that adding an empty cell acted like adding 0, and
multiplying by an empty cell acted like multiplying by 1. This proved to
be surprising.