Impletemnt `#:prcsn`
This adds a concept of "rolling formats" and more formalizes the idea of
formats. `#:prcsn` is designed to set the precision value of the next
row in a way that overwrites the general rule designated by `#:fmt`. The
current "rolling format" can be ended either by `#:prcsn reset` or by
inserting a blank line. These two methods are identical but for the
blank line's dual purpose of headers and footers from the body. This
presently encourages adding blank lines within the body (a questionable
but supported practice) and so the more specific `#:prcsn reset` was
added as quick solution.