Suplex all of our memory stuff into mem.h
Hopefully I got it all. The point was so that we could hopefully have
the function in charge of printing our memory usage be more accurate.
Plus, in the future, it may make it easier to shift away from using
malloc as much.

I say "suplex" because there was little thought in the procedure, and I
can't say I'm any happier with the file layout.