Avoid reducing nodes in place.
It's unclear that I will see any performance benefits or memory usage
benefits from this, but it does make me just a little bit happier to
know that I don't re-parse the macro's body every time it's used. The
trade off, for now, is that I may need to allocate nodes while reducing
a list node, but I expect that will usually result in a net reduction in
nodes allocated in most of my use cases.