Add relative xeno refs and `cell`, `col`, and `row` functions
Relative xeno refs allow me to ask "What's in this same location, but in
this other file?" I originally forbade that, but it came in handy now.

`col/0` returns the column index of the current cell, and `row/0`
returns the row index of the current cell. `cell/2` takes a column index
and row index and returns the value of the cell, if any, at that
location. `cell/3` works similarly, but the first parameter is a string
and will be interpreted as a xeno reference. These functions allow me to
do more complex relative math. The previous relative notation will be
kept as a shorthand for simple relative math.