Add round and trunc functions; massively refactor functions
The goal was to simplify the process of adding new functions to the
code. The success was marginal, but I did manage to localize the changes
to two places: near the top of the file where the function enum and
specifications are defined and in the implementation switch statement.
That should help keep things consistant and cut down on boilerplate.
We'll see if this ever bites me in the ass.

Some stylistic changes got caught up in the process.